move scale's init/parse/common constants & interfaces to `src/compile/scale/init,parse,scale`
